refactor folders n things
