+++ .vscode/settings.json
@@ -1,1 +1,2 @@
-{}
+{
+}
